BS+mergesort
[0,4] should be merged only after [0,2]and [3,4] is merged.
In this aspect, [0,2] should be merged in the process after [0,1], [0,2] is done.
when a, b is a>=b, MS() returns a(or b) it self- which means recursion is over- 'then' it is ready to run next line.
Thus,
    MS(L,first,m)
    MS(L,m+1,last)
being written before merge() is the very function -sort of a device- that the bigger scale of merging process can always be done after their small-parts-merging process.
reaching line merge(L,first,last) means that L[first,m] and L[m+1,last] is already sorted respectively.