It is important to figure out when the necessary info needed to define accum[v] is ready.
In this problem,
The necessary info to define accum[v] is always perfectly defined before the attempt to define accum[v]. It is because the raw input is given that way(in a numeric order, so defining all accum[v] is always ready to be finished because the info needed is always define before.

What if the info needed is not ready?

The goal is to collect all infos that affects current vertax(v).
The number of infos related to current v is the num of edge (v0 ->v)
Which means, when v0 has finished to search its max dur(and found it),
it is then ready to pass over its info to those that need it.

How can I know when a vertax is ready to pass down its info?
It is determined when its indegree is 0

A vertax with indegree 0 means there is no info to collect in order to complete info of itself.

when a vertex with indegree 0 passes down its info to v2,
decrease indegree[v2] once to inform v2 that its amount of effort to search its valid info
has decreased once.

So in this case, list indegree, accum is needed.
list accum  of v2 is renewed each time it receives info(v->v2, such v that its indegree is or has turned to 0 )

In another aspect,
v can collect its info with accum[v2] when v->v2
v2 is the info NEEDED to finish v. (make sure to set the heading edge according to it)

but accum[v2] in that instance might not have finshed to renew itself.
How will I know then accum[v2] is ready?
It can be determined by checking indegree[v2]==0
What if indegree[v2]!=0? It means it is of no use to bring accum[v2].

This is why an order of finishing accum[v] should always be placed according to the time line indegree[v] turns 0.