01knapsack errors
In cp function, in case backpack is not fully packed but no goods are left, return ans should've been when for loop is over.

When sorting cp, reverse

The process of determining whethere to implement a case or not should inspect values, not weights.

p=0
p+=sum([])
has no errors

Renewing mx is only and always done when current i seems suitable to be packed