radix ranking, qs pivot
C = [0]*10 counts elements when index of C is F[el,d].
if C = [0,2,0,1,1,3],
2 elements of C[1]'s rank is 2 because it is joint ranking.
1 element of C[3] is ranked 3, so it C[3]+=C[:3]. To accumulate rankingm the process of C[i+1]+=C[i] for i in range(1,10) should be done.

To sort L based on C's rank, firstly find each element's location which is f[el,d].
The sorted list for L is defined B, so in B[C[f(el,d)]-1], el should be laid. For more explanation, -1 in B[C[f(el,d)]-1]  is done because list B's index starts with 0 so rank1 should go inside B[1-1]=B[0].
For elements having joint rankings not to be overlapped in same B[i], C[f(el,d)]-=1 should be done.

It should be noted that L is sorted for d-1. To maintain its ranks for d-1and sort for d, the process of storation for elements in B should be done in reversed L. () Because in same C[i], bigger number for d-1 should be ranked higher.( The latter C[i] is decreasing since C[i]-=1.)

The process of QS is spliting L into half where the elements of first half is not bigger than all elements of latter half, and the elements of latter half is not smaller than all elements of previous half.