MoM: find median of 6 numvers.
1. Run a tournament for L[0:5] and place round1 winners' index in R[0], R[1] each, and place round 2 winners' index in R[2].

2. Erase the winner and fill its space with L[4], the unearned winner.
L[R[2]]= L[4]

3. find the winner again. Now the 1st and 2nd winner is found; its order is unknown; so the winner of remaining 3 is the median.

4. R[2]'s location should be detected: What round1 did R[2] participate in?
If its index is 0 or 1, it was the element that won R[0], and if it is 1 or 2, R[1]
Therefore, by index//2, its place that held round1 can be distinguished.

5. R[2]'s location should be detected: Where were R[2] placed based on round1: left or right?
If the index is 0 or 2, it means R[2] was in leftside: both 0,2 %2==0
Elif the index is 1 or 3, it means R[2] was in rightside:
both 1,3 %2 ==1.