3363 coin sol
It is reasonable that the 4 num in the front and 4 in the back has opposite sign, whether the real sign is > or <.

To use this given info, make each num(from 1~12) (3 * 2) storing spaces that hold sign of each 3 line.

fill all spaces with '='.

for each num, it has two big storehouse, ['=','=','='] , ['=','=','='] each.

4 num in the front and 4 in the back has opposite sign, whether the real sign is > or <.

to use the current num's location to define its possibility, put > in the 1st storage room (< in the 2nd storage room) for 4 numbers given from the left, and vise versa for the right 4 numbers.

When the storing process is done, check if it matches 3 given sign.
It it matches, it can be the false coin.
Its weight can be distinguished whether the match has occured in the 1st storage or 2nd.

when current line's sign is '=', no numer in that line can be the false coin.