Why's Java called OOP, constr. overloading, this.var, this()
1. Java: Object Oriented Programming?
Java  depicts, describes real world's Object into concept of software Object well.

2.why const overloading?
If max parameter is 3 but I only want to change one or more and leave the other to its initial value, new constructer is made that has one or two parameter.

3. this.val?
defining val equals to defining field.
Object val should be defined outside of  braces.

4. this()?
to shorten code,
in case of num 2,
the new constructer's inside can be like this:

Car (String a, String b){
this(a, b, init_value)
}

Car (String a, Strinig b, int c)
{
this.a=a;
this.b=b;
this.c=c;
}